Rename packages to follow convention for Go
