+++ dgraph/cmd/graphql/e2e/directives/dgraph_directives_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package complex
+package directives
 
 import (
 	"io/ioutil"
