+++ dgraph/cmd/graphql/e2e/normal/normal_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package complex
+package normal
 
 import (
 	"io/ioutil"
